refactor: change app structure
Shift to `cmd` and `pkg` structure, influenced by Kat Zien talk
"How Do You Structure Your Go Apps?" at GopherCon UK 2018. Added an
architecture description, influenced by Domain Driven Design and
Hexagonal Architecture.
