+++ cmd/server/main_test.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"fmt"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
+	"strconv"
 	"strings"
 	"testing"
 )
@@ -34,12 +34,12 @@ func TestEndpoints(t *testing.T) {
 		{name: "no-path", url: "", status: http.StatusOK, content: "<h1>Learn to Create Websites</h1>"},
 		{name: "index", url: "index.html", status: http.StatusOK, content: "<h1>Learn to Create Websites</h1>"},
 		{name: "about", url: "about.html", status: http.StatusOK, content: "<h1>About</h1>"},
-		{name: "favicon", url: "favicon.ico", status: http.StatusOK},
+		// {name: "favicon", url: "favicon.ico", status: http.StatusOK},
 	}
 
 	// IMPORTANT: comment/uncomment to change where the app is running
-	// prefix := "http://localhost:" + strconv.Itoa(srv.cfg.port)
-	prefix := fmt.Sprintf("https://%s.appspot.com", srv.cfg.projectID)
+	prefix := "http://localhost:" + strconv.Itoa(srv.cfg.port)
+	// prefix := fmt.Sprintf("https://%s.appspot.com", srv.cfg.projectID)
 
 	for _, tc := range tests {
 		endpoint := tc.url
