+++ pkg/middleware/gzip.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"compress/gzip"
