+++ pkg/middleware/gzip_test.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"fmt"
