+++ pkg/middleware/logReqResp.go
@@ -1,4 +1,4 @@
-package main
+package middleware
 
 import (
 	"log"
